Split editor type config tabs to separate constants (#601)
