+++ src/shared/constants/common.ts
@@ -43,3 +43,37 @@ export const URL_ACTION_PARAMS_PREFIX = ACTION_PARAM_PREFIX;
 export const DEFAULT_CHART_LINES_LIMIT = 100;
 
 export const DEFAULT_DATE_FORMAT = 'DD.MM.YYYY';
+
+export const WIZARD_CHART_NODE = {
+    graph_wizard_node: 'statface_graph',
+    table_wizard_node: 'table',
+    metric_wizard_node: 'statface_metric',
+};
+
+export const QL_CHART_NODE = {
+    graph_sql_node: 'statface_graph',
+    graph_ql_node: 'statface_graph',
+    timeseries_ql_node: 'statface_graph',
+    table_sql_node: 'table',
+    table_ql_node: 'table',
+};
+
+export const EDITOR_CHART_NODE = {
+    graph_node: 'statface_graph',
+    graph_billing_node: 'statface_graph',
+    table_node: 'table',
+    text_node: 'statface_text',
+    metric_node: 'statface_metric',
+    metric_sql_node: 'statface_metric',
+    metric_ql_node: 'statface_metric',
+    map_node: 'statface_map',
+    markup_node: 'config',
+    markdown_node: '',
+};
+
+export const EDITOR_TYPE_CONFIG_TABS = {
+    ...WIZARD_CHART_NODE,
+    ...QL_CHART_NODE,
+    ...EDITOR_CHART_NODE,
+    module: '',
+};
